requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.1
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
django-seed==0.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Faker==8.2.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
